+++ R/file_metadata.R
@@ -7,6 +7,8 @@
 #'   file."
 #' @param sys_version sourceSystemVersion, defined as "The version of the
 #'   sourceSystem"
+#' @param file_oid fileOID parameter, defined as "A unique identifier for this
+#'   file."
 #' @param version Dataset JSON schema version being used
 #'
 #' @return file_metadata object
@@ -26,17 +28,17 @@
 #' file_meta_updated <- set_file_oid(file_meta, "/some/path")
 #' file_meta_updated <- set_originator(file_meta_updated, "Some Org")
 #' file_meta_updated <- set_source_system(file_meta_updated, "source system", "1.0")
-file_metadata <- function(originator="NA", sys = "NA", sys_version = "NA", version = "1.0.0") {
+file_metadata <- function(originator=NULL, sys = NULL, sys_version = NULL, file_oid = NULL, version = "1.0.0") {
 
   if (!(version %in% c("1.0.0"))) {
     stop("Unsupported version specified - currently only version 1.0.0 is supported", call.=FALSE)
   }
 
   x <- list(
-    "creationDateTime"= get_datetime(),
+    "creationDateTime"= character(),
     "datasetJSONVersion"= version,
-    "fileOID" = character(),
-    "asOfDateTime" = character(),
+    "fileOID" = file_oid,
+    "asOfDateTime" = NULL, # Not sure we want this to exist?
     "originator" = originator,
     "sourceSystem" = sys,
     "sourceSystemVersion" = sys_version
@@ -59,10 +61,20 @@ get_datetime <- function() {
   format(Sys.time(), "%Y-%m-%dT%H:%M:%S")
 }
 
-#' Set source system information
+#' File Metadata Setters
 #'
-#' Set information about the source system used to generate the Dataset JSON
-#' object.
+#' Set information about the file and source system used to generate the Dataset
+#' JSON object.
+#'
+#' @details
+#'
+#' The fileOID parameter should be structured following description outlined in
+#' the ODM V2.0 specification. "FileOIDs should be universally unique if at all
+#' possible. One way to ensure this is to prefix every FileOID with an internet
+#' domain name owned by the creator of the ODM file or database (followed by a
+#' forward slash, "/"). For example,
+#' FileOID="BestPharmaceuticals.com/Study5894/1" might be a good way to denote
+#' the first file in a series for study 5894 from Best Pharmaceuticals."
 #'
 #' @param x datasetjson object
 #' @param sys sourceSystem parameter, defined as "The computer system or