+++ NEWS.md
@@ -1,3 +1,12 @@
-# datasetjson 0.0.0.9999
+# datasetjson 0.1.0
 
-Intial development version of datasetjson, introducing core objects, readers and writers.
+- Capability to read and validate Dataset JSON files from URLs has been added (#8)
+- Remove autoset of fileOID using output path (#3)
+- Don't auto-populate optional attributes with NA (#16)
+- Push dependency versions back (#18)
+- Default `pretty` parameter on `write_dataset_json()` to false (#20)
+
+# datasetjson 0.0.1
+
+Initial development version of datasetjson, introducing core objects, readers and writers.
+